Network Scanner
Keyo Scan is a powerful scanning tool developed using Scapy Python library that helps in scanning and analyzing networks and connected devices, discovering systems and opening ports in the network. The tool provides multiple options to the user to enable him to perform accurate and effective scans on the target network.